refactor(storage): extract Blazegraph N-Quads encoder + centralize SPARQL charset
Addresses round-4 review (both structural):
- New packages/storage/src/adapters/blazegraph-nquads.ts holds the ASCII-safe
  N-Quads wire encoder (toBlazegraphAsciiSafeNQuads + escapeUtf16CodeUnits),
  a focused boundary separate from the adapter's transport/query/mutation code.
  blazegraph.ts imports it; quadToBlazegraphNQuad stays as the write-path
  serializer. Byte behavior unchanged.
- New packages/storage/src/adapters/sparql-content-types.ts centralizes
  SPARQL_QUERY_CONTENT_TYPE / SPARQL_UPDATE_CONTENT_TYPE (both charset=utf-8),
  used by the blazegraph + sparql-http adapters so future request paths can't
  drift back to a charset-less Content-Type.

Pure refactor — no wire-behavior change; storage suite 343 passed.
